+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "vscode-hugo-runner" extension will be documented in 
 ## vNext
 
 - Added: WebView options to build drafts, future, and expired content
+- Removed: `hugoRunner.showDrafts` setting in favour of the WebView option
 
 ## v0.1.2 2024-07-02